--- app/src/main/java/com/mypet/FactoryPattern/TestClass.java
@@ -1,9 +1,0 @@
-package com.mypet.FactoryPattern;
-
-public class TestClass {
-
-
-    public static void main(String[] args) {
-        IPet pet1 = new Dog();
-    }
-}
