+++ app/src/main/java/com/mypet/PetFactory/Hamster.java
@@ -1,13 +1,12 @@
-package com.mypet.FactoryPattern;
+package com.mypet.PetFactory;
 
 import android.util.Log;
 
 import static android.content.ContentValues.TAG;
 
-public class Cat implements IPet {
-
+public class Hamster implements IPet {
     @Override
     public void message() {
-        Log.d(TAG, "createPet:Cat is created ");
+        Log.d(TAG, "createPet:Hamster is created ");
     }
 }
