+++ app/src/main/java/com/mypet/PetFactory/IPet.java
@@ -1,4 +1,4 @@
-package com.mypet.FactoryPattern;
+package com.mypet.PetFactory;
 
 public interface IPet {
     void message();
