+++ app/src/main/java/com/mypet/PetFactory/PetTypeFactory.java
@@ -1,4 +1,4 @@
-package com.mypet.FactoryPattern;
+package com.mypet.PetFactory;
 
 import com.mypet.Enums.AnimalType;
